# docs/website/references/http/openapi.json
{
    "components": {
...
        "responses": {
            "error": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/error"
                        }
                    }
                },
                "description": "error"
            },
            "success": {
                "description": "ok"
            }
        },
        "schemas": {
...
            "delete_replicator_params": {
                "properties": {
                    "Collections": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "ID": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "error": {
                "properties": {
                    "error": {}
                },
                "type": "object"
            },
...
            "peer_info": {
                "properties": {
                    "Addresses": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "ID": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "replicator": {
                "properties": {
                    "Addresses": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "CollectionIDs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "ID": {
                        "type": "string"
                    },
                    "LastStatusChange": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "Status": {
                        "maximum": 255,
                        "minimum": 0,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "set_replicator_params": {
                "properties": {
                    "Addresses": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "Collections": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerToken": {
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
...
    "info": {
        "title": "DefraDB API",
        "version": "0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/p2p/active-peers": {
            "get": {
                "operationId": "active_peers",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Connected peers"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
        "/p2p/collections": {
            "delete": {
                "description": "Remove peer collections",
                "operationId": "peer_collections_remove",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            },
            "get": {
                "description": "List peer collections",
                "operationId": "peer_collections_list",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Peer collections"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            },
            "post": {
                "description": "Add peer collections",
                "operationId": "peer_collections_add",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
        "/p2p/collections/sync-versions": {
            "post": {
                "description": "Synchronize collection versions to the local node",
                "operationId": "peer_sync_collection_versions",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "timeout": {
                                        "type": "string"
                                    },
                                    "versionIDs": {
                                        "items": {
                                            "type": "string"
                                        },
                                        "type": "array"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Collection synchronization completed successfully"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "500": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
        "/p2p/connect": {
            "post": {
                "operationId": "connect",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
        "/p2p/documents": {
            "delete": {
                "description": "Remove peer documents",
                "operationId": "peer_documents_remove",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            },
            "get": {
                "description": "List peer documents",
                "operationId": "peer_documents_list",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "type": "string"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Peer documents"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            },
            "post": {
                "description": "Add peer documents",
                "operationId": "peer_documents_add",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
        "/p2p/documents/sync": {
            "post": {
                "description": "Synchronize documents from the network",
                "operationId": "peer_sync_documents",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "collectionName": {
                                        "type": "string"
                                    },
                                    "docIDs": {
                                        "items": {
                                            "type": "string"
                                        },
                                        "type": "array"
                                    },
                                    "timeout": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Document sync completed successfully"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "500": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
        "/p2p/info": {
            "get": {
                "operationId": "peer_info",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/peer_info"
                                }
                            }
                        },
                        "description": "Peer network info"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
        "/p2p/replicators": {
            "delete": {
                "description": "Delete peer replicators",
                "operationId": "peer_replicator_delete",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/delete_replicator_params"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            },
            "get": {
                "description": "List peer replicators",
                "operationId": "peer_replicator_list",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/replicator"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Replicators"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            },
            "post": {
                "description": "Add peer replicators",
                "operationId": "peer_replicator_set",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/set_replicator_params"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "p2p"
                ]
            }
        },
...
    },
    "servers": [
        {
            "description": "Local DefraDB instance",
            "url": "/api/v0"
        }
    ],
...
}