# http/handler_p2p.go
// Copyright 2023 Democratized Data Foundation
//
// Use of this software is governed by the Business Source License
// included in the file licenses/BSL.txt.
//
// As of the Change Date specified in that file, in accordance with
// the Business Source License, use of this software will be governed
// by the Apache License, Version 2.0, included in the file
// licenses/APL.txt.

package http

import (
	"context"
	"net/http"
	"time"

	"github.com/getkin/kin-openapi/openapi3"
)

type p2pHandler struct{}

func (h *p2pHandler) PeerInfo(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)
	addresses, err := db.PeerInfo()
	if err != nil {
		responseJSON(rw, http.StatusInternalServerError, errorResponse{err})
		return
	}
	responseJSON(rw, http.StatusOK, addresses)
}

func (h *p2pHandler) ActivePeers(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)
	peers, err := db.ActivePeers(req.Context())
	if err != nil {
		responseJSON(rw, http.StatusInternalServerError, errorResponse{err})
		return
	}
	responseJSON(rw, http.StatusOK, peers)
}

func (h *p2pHandler) Connect(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var resp []string
	if err := requestJSON(req, &resp); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	err := db.Connect(req.Context(), resp)
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) SetReplicator(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var rep SetReplicatorParams
	if err := requestJSON(req, &rep); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	err := db.SetReplicator(req.Context(), rep.Addresses, rep.Collections...)
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) DeleteReplicator(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var rep DeleteReplicatorParams
	if err := requestJSON(req, &rep); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	err := db.DeleteReplicator(req.Context(), rep.ID, rep.Collections...)
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) GetAllReplicators(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	reps, err := db.GetAllReplicators(req.Context())
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	responseJSON(rw, http.StatusOK, reps)
}

func (h *p2pHandler) AddP2PCollections(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var collectionIDs []string
	if err := requestJSON(req, &collectionIDs); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	err := db.AddP2PCollections(req.Context(), collectionIDs...)
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) RemoveP2PCollections(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var collectionIDs []string
	if err := requestJSON(req, &collectionIDs); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	err := db.RemoveP2PCollections(req.Context(), collectionIDs...)
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) GetAllP2PCollections(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	cols, err := db.GetAllP2PCollections(req.Context())
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	responseJSON(rw, http.StatusOK, cols)
}

func (h *p2pHandler) AddP2PDocuments(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var docIDs []string
	if err := requestJSON(req, &docIDs); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	err := db.AddP2PDocuments(req.Context(), docIDs...)
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) RemoveP2PDocuments(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var docIDs []string
	if err := requestJSON(req, &docIDs); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	err := db.RemoveP2PDocuments(req.Context(), docIDs...)
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) GetAllP2PDocuments(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	docIDs, err := db.GetAllP2PDocuments(req.Context())
	if err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}
	responseJSON(rw, http.StatusOK, docIDs)
}

func (h *p2pHandler) SyncDocuments(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var reqBody struct {
		CollectionName string   `json:"collectionName"`
		DocIDs         []string `json:"docIDs"`
		Timeout        string   `json:"timeout"`
	}

	if err := requestJSON(req, &reqBody); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}

	ctx := req.Context()
	if reqBody.Timeout != "" {
		timeout, err := time.ParseDuration(reqBody.Timeout)
		if err != nil {
			responseJSON(rw, http.StatusBadRequest, errorResponse{err})
			return
		}
		var cancel context.CancelFunc
		ctx, cancel = context.WithTimeout(ctx, timeout)
		defer cancel()
	}

	err := db.SyncDocuments(ctx, reqBody.CollectionName, reqBody.DocIDs)
	if err != nil {
		responseJSON(rw, http.StatusInternalServerError, errorResponse{err})
		return
	}

	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) SyncCollectionVersions(rw http.ResponseWriter, req *http.Request) {
	db := mustGetContextClientDB(req)

	var reqBody struct {
		VersionIDs []string `json:"versionIDs"`
		Timeout    string   `json:"timeout"`
	}

	if err := requestJSON(req, &reqBody); err != nil {
		responseJSON(rw, http.StatusBadRequest, errorResponse{err})
		return
	}

	ctx := req.Context()
	if reqBody.Timeout != "" {
		timeout, err := time.ParseDuration(reqBody.Timeout)
		if err != nil {
			responseJSON(rw, http.StatusBadRequest, errorResponse{err})
			return
		}
		var cancel context.CancelFunc
		ctx, cancel = context.WithTimeout(ctx, timeout)
		defer cancel()
	}

	err := db.SyncCollectionVersions(ctx, reqBody.VersionIDs...)
	if err != nil {
		responseJSON(rw, http.StatusInternalServerError, errorResponse{err})
		return
	}

	rw.WriteHeader(http.StatusOK)
}

func (h *p2pHandler) bindRoutes(router *Router) {
	successResponse := &openapi3.ResponseRef{
		Ref: "#/components/responses/success",
	}
	errorResponse := &openapi3.ResponseRef{
		Ref: "#/components/responses/error",
	}
	peerInfoSchema := &openapi3.SchemaRef{
		Ref: "#/components/schemas/peer_info",
	}
	replicatorSchema := &openapi3.SchemaRef{
		Ref: "#/components/schemas/replicator",
	}
	setReplicatorParamsSchema := &openapi3.SchemaRef{
		Ref: "#/components/schemas/set_replicator_params",
	}
	deleteReplicatorParamsSchema := &openapi3.SchemaRef{
		Ref: "#/components/schemas/delete_replicator_params",
	}

	peerInfoResponse := openapi3.NewResponse().
		WithDescription("Peer network info").
		WithContent(openapi3.NewContentWithJSONSchemaRef(peerInfoSchema))

	peerInfo := openapi3.NewOperation()
	peerInfo.OperationID = "peer_info"
	peerInfo.Tags = []string{"p2p"}
	peerInfo.AddResponse(200, peerInfoResponse)
	peerInfo.Responses.Set("400", errorResponse)

	activePeersSchema := openapi3.NewArraySchema().
		WithItems(openapi3.NewStringSchema())

	activePeersResponse := openapi3.NewResponse().
		WithDescription("Connected peers").
		WithContent(openapi3.NewContentWithJSONSchema(activePeersSchema))

	activePeers := openapi3.NewOperation()
	activePeers.OperationID = "active_peers"
	activePeers.Tags = []string{"p2p"}
	activePeers.AddResponse(200, activePeersResponse)
	activePeers.Responses.Set("400", errorResponse)

	connect := openapi3.NewOperation()
	connect.OperationID = "connect"
	connect.Tags = []string{"p2p"}
	connect.Responses = openapi3.NewResponses()
	connect.Responses.Set("200", successResponse)
	connect.Responses.Set("400", errorResponse)

	getReplicatorsSchema := openapi3.NewArraySchema()
	getReplicatorsSchema.Items = replicatorSchema
	getReplicatorsResponse := openapi3.NewResponse().
		WithDescription("Replicators").
		WithContent(openapi3.NewContentWithJSONSchema(getReplicatorsSchema))

	getReplicators := openapi3.NewOperation()
	getReplicators.Description = "List peer replicators"
	getReplicators.OperationID = "peer_replicator_list"
	getReplicators.Tags = []string{"p2p"}
	getReplicators.AddResponse(200, getReplicatorsResponse)
	getReplicators.Responses.Set("400", errorResponse)

	setReplicatorRequest := openapi3.NewRequestBody().
		WithRequired(true).
		WithContent(openapi3.NewContentWithJSONSchemaRef(setReplicatorParamsSchema))

	setReplicator := openapi3.NewOperation()
	setReplicator.Description = "Add peer replicators"
	setReplicator.OperationID = "peer_replicator_set"
	setReplicator.Tags = []string{"p2p"}
	setReplicator.RequestBody = &openapi3.RequestBodyRef{
		Value: setReplicatorRequest,
	}
	setReplicator.Responses = openapi3.NewResponses()
	setReplicator.Responses.Set("200", successResponse)
	setReplicator.Responses.Set("400", errorResponse)

	deleteReplicatorRequest := openapi3.NewRequestBody().
		WithRequired(true).
		WithContent(openapi3.NewContentWithJSONSchemaRef(deleteReplicatorParamsSchema))

	deleteReplicator := openapi3.NewOperation()
	deleteReplicator.Description = "Delete peer replicators"
	deleteReplicator.OperationID = "peer_replicator_delete"
	deleteReplicator.Tags = []string{"p2p"}
	deleteReplicator.RequestBody = &openapi3.RequestBodyRef{
		Value: deleteReplicatorRequest,
	}
	deleteReplicator.Responses = openapi3.NewResponses()
	deleteReplicator.Responses.Set("200", successResponse)
	deleteReplicator.Responses.Set("400", errorResponse)

	peerCollectionsSchema := openapi3.NewArraySchema().
		WithItems(openapi3.NewStringSchema())

	peerCollectionRequest := openapi3.NewRequestBody().
		WithRequired(true).
		WithContent(openapi3.NewContentWithJSONSchema(peerCollectionsSchema))

	getPeerCollectionsResponse := openapi3.NewResponse().
		WithDescription("Peer collections").
		WithContent(openapi3.NewContentWithJSONSchema(peerCollectionsSchema))

	getPeerCollections := openapi3.NewOperation()
	getPeerCollections.Description = "List peer collections"
	getPeerCollections.OperationID = "peer_collections_list"
	getPeerCollections.Tags = []string{"p2p"}
	getPeerCollections.AddResponse(200, getPeerCollectionsResponse)
	getPeerCollections.Responses.Set("400", errorResponse)

	addPeerCollections := openapi3.NewOperation()
	addPeerCollections.Description = "Add peer collections"
	addPeerCollections.OperationID = "peer_collections_add"
	addPeerCollections.Tags = []string{"p2p"}
	addPeerCollections.RequestBody = &openapi3.RequestBodyRef{
		Value: peerCollectionRequest,
	}
	addPeerCollections.Responses = openapi3.NewResponses()
	addPeerCollections.Responses.Set("200", successResponse)
	addPeerCollections.Responses.Set("400", errorResponse)

	removePeerCollections := openapi3.NewOperation()
	removePeerCollections.Description = "Remove peer collections"
	removePeerCollections.OperationID = "peer_collections_remove"
	removePeerCollections.Tags = []string{"p2p"}
	removePeerCollections.RequestBody = &openapi3.RequestBodyRef{
		Value: peerCollectionRequest,
	}
	removePeerCollections.Responses = openapi3.NewResponses()
	removePeerCollections.Responses.Set("200", successResponse)
	removePeerCollections.Responses.Set("400", errorResponse)

	peerDocumentsSchema := openapi3.NewArraySchema().
		WithItems(openapi3.NewStringSchema())

	peerDocumentRequest := openapi3.NewRequestBody().
		WithRequired(true).
		WithContent(openapi3.NewContentWithJSONSchema(peerDocumentsSchema))

	getPeerDocumentsResponse := openapi3.NewResponse().
		WithDescription("Peer documents").
		WithContent(openapi3.NewContentWithJSONSchema(peerDocumentsSchema))

	getPeerDocuments := openapi3.NewOperation()
	getPeerDocuments.Description = "List peer documents"
	getPeerDocuments.OperationID = "peer_documents_list"
	getPeerDocuments.Tags = []string{"p2p"}
	getPeerDocuments.AddResponse(200, getPeerDocumentsResponse)
	getPeerDocuments.Responses.Set("400", errorResponse)

	addPeerDocuments := openapi3.NewOperation()
	addPeerDocuments.Description = "Add peer documents"
	addPeerDocuments.OperationID = "peer_documents_add"
	addPeerDocuments.Tags = []string{"p2p"}
	addPeerDocuments.RequestBody = &openapi3.RequestBodyRef{
		Value: peerDocumentRequest,
	}
	addPeerDocuments.Responses = openapi3.NewResponses()
	addPeerDocuments.Responses.Set("200", successResponse)
	addPeerDocuments.Responses.Set("400", errorResponse)

	removePeerDocuments := openapi3.NewOperation()
	removePeerDocuments.Description = "Remove peer documents"
	removePeerDocuments.OperationID = "peer_documents_remove"
	removePeerDocuments.Tags = []string{"p2p"}
	removePeerDocuments.RequestBody = &openapi3.RequestBodyRef{
		Value: peerDocumentRequest,
	}
	removePeerDocuments.Responses = openapi3.NewResponses()
	removePeerDocuments.Responses.Set("200", successResponse)
	removePeerDocuments.Responses.Set("400", errorResponse)

	syncDocumentsRequestSchema := openapi3.NewObjectSchema().
		WithProperty("collectionName", openapi3.NewStringSchema()).
		WithProperty("docIDs", openapi3.NewArraySchema().WithItems(openapi3.NewStringSchema())).
		WithProperty("timeout", openapi3.NewStringSchema())

	syncDocumentsRequest := openapi3.NewRequestBody().
		WithRequired(true).
		WithContent(openapi3.NewContentWithJSONSchema(syncDocumentsRequestSchema))

	syncDocumentsResponse := openapi3.NewResponse().
		WithDescription("Document sync completed successfully")

	syncDocuments := openapi3.NewOperation()
	syncDocuments.Description = "Synchronize documents from the network"
	syncDocuments.OperationID = "peer_sync_documents"
	syncDocuments.Tags = []string{"p2p"}
	syncDocuments.RequestBody = &openapi3.RequestBodyRef{
		Value: syncDocumentsRequest,
	}
	syncDocuments.Responses = openapi3.NewResponses()
	syncDocuments.Responses.Set("200", &openapi3.ResponseRef{Value: syncDocumentsResponse})
	syncDocuments.Responses.Set("400", errorResponse)
	syncDocuments.Responses.Set("500", errorResponse)

	syncCollectionVersionsRequestSchema := openapi3.NewObjectSchema().
		WithProperty("versionIDs", openapi3.NewArraySchema().WithItems(openapi3.NewStringSchema())).
		WithProperty("timeout", openapi3.NewStringSchema())

	syncCollectionVersionsRequest := openapi3.NewRequestBody().
		WithRequired(true).
		WithContent(openapi3.NewContentWithJSONSchema(syncCollectionVersionsRequestSchema))

	syncCollectionVersionsResponse := openapi3.NewResponse().
		WithDescription("Collection synchronization completed successfully")

	syncCollectionVersions := openapi3.NewOperation()
	syncCollectionVersions.Description = "Synchronize collection versions to the local node"
	syncCollectionVersions.OperationID = "peer_sync_collection_versions"
	syncCollectionVersions.Tags = []string{"p2p"}
	syncCollectionVersions.RequestBody = &openapi3.RequestBodyRef{
		Value: syncCollectionVersionsRequest,
	}
	syncCollectionVersions.Responses = openapi3.NewResponses()
	syncCollectionVersions.Responses.Set("200", &openapi3.ResponseRef{Value: syncCollectionVersionsResponse})
	syncCollectionVersions.Responses.Set("400", errorResponse)
	syncCollectionVersions.Responses.Set("500", errorResponse)

	router.AddRoute("/p2p/info", http.MethodGet, peerInfo, h.PeerInfo)
	router.AddRoute("/p2p/active-peers", http.MethodGet, activePeers, h.ActivePeers)
	router.AddRoute("/p2p/connect", http.MethodPost, connect, h.Connect)
	router.AddRoute("/p2p/replicators", http.MethodGet, getReplicators, h.GetAllReplicators)
	router.AddRoute("/p2p/replicators", http.MethodPost, setReplicator, h.SetReplicator)
	router.AddRoute("/p2p/replicators", http.MethodDelete, deleteReplicator, h.DeleteReplicator)
	router.AddRoute("/p2p/collections", http.MethodGet, getPeerCollections, h.GetAllP2PCollections)
	router.AddRoute("/p2p/collections", http.MethodPost, addPeerCollections, h.AddP2PCollections)
	router.AddRoute("/p2p/collections", http.MethodDelete, removePeerCollections, h.RemoveP2PCollections)
	router.AddRoute("/p2p/collections/sync-versions", http.MethodPost, syncCollectionVersions, h.SyncCollectionVersions)
	router.AddRoute("/p2p/documents", http.MethodGet, getPeerDocuments, h.GetAllP2PDocuments)
	router.AddRoute("/p2p/documents", http.MethodPost, addPeerDocuments, h.AddP2PDocuments)
	router.AddRoute("/p2p/documents", http.MethodDelete, removePeerDocuments, h.RemoveP2PDocuments)
	router.AddRoute("/p2p/documents/sync", http.MethodPost, syncDocuments, h.SyncDocuments)
}